requirements.txt
# SuperDeploy CLI Dependencies
click>=8.1.0
rich>=13.0.0
rich-click>=1.7.0
inquirer>=3.1.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
jinja2>=3.1.0
requests>=2.31.0
python-terraform>=0.10.1
paramiko>=3.0.0

# Google Cloud
google-cloud-resource-manager>=1.12.0

# PERFORMANCE: Ansible + Mitogen in venv (no system dependency!)
# Note: Ansible 9.x+ required for community.docker 5.0.0 compatibility (Python 3.10+)
ansible>=9.0.0,<13.0.0
ansible-runner>=2.3.0
mitogen>=0.3.4

# COLLECTIONS: community.docker with docker_compose_v2 support
# Note: Will be auto-installed with ansible, but specify version for docker_compose_v2
# Install manually: ansible-galaxy collection install community.docker:>=5.0.0

# DASHBOARD: Local web UI for secrets management
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.13.0
